Fixing the script for cleaning workstation

+++ clean_ws.sh
@@ -1,10 +1,16 @@
 #!/bin/bash
 
+# Make sure only root can run our script
+if [ "$(id -u)" != "0" ]; then
+   echo "This script must be run as root" 1>&2
+   exit 1
+fi
+
 echo "Remove the kubernetes packages and other old packages"
 apt remove -y kubectl kubernetes filebeat quagga
 
 echo "Remove the unnecessary repos from system"
-rm -f /etc/apt/sources.list.d/*?????shift*
+rm -fv /etc/apt/sources.list.d/*shift*
 
 echo "Stop and remove the kubernetes services"
 for service in kubelet kube-proxy; do
@@ -23,10 +29,7 @@ apt-key del 574E4513ACDFDF63
 
 echo "Delete tls keys"
 rm -rf /etc/?????shift/tls
-rm -rf /etc/docker/certs.d/*?????shift*
-
-echo "Delete the puppet alias config file"
-[ -f /etc/profile.d/puppet-aliases.sh ] && rm -f /etc/profile.d/puppet-aliases.sh
+rm -rf /etc/docker/certs.d/*shift*
 
 echo "Fix the resolv.conf file"
 sed 's/^\(.*\) .*svc\.ru-nsk-1\.k8s\(.*$\)/\1 \2/' -i /etc/resolv.conf
@@ -44,8 +47,9 @@ rm -f /etc/cni/net.d/*weave*
 
 echo "Clean the docker from ALL images, containers, volumes etc"
 service docker stop
-rm -f /var/lib/docker/*
+rm -rf /var/lib/docker/*
 service docker start
 
-echo "Delete the packages were automatically installed and are no longer required"
-apt autoremove
+echo -e "If you are wanna clean your system you can start the command below to:
+'delete the packages were automatically installed and are no longer required'\n
+sudo apt autoremove -y"
